Revert "Moved components to shared"
This reverts commit 219e86bd2dabc032b2572cba3d71639c19bea401.
