+++ front/components/assistant/conversation/GenerationContextProvider.tsx
