+++ front/components/assistant/conversation/input_bar/InputBarCitations.tsx
