+++ front/components/assistant/conversation/input_bar/editor/useAssistantSuggestions.ts
@@ -1,7 +1,11 @@
-import type { LightAgentConfigurationType } from "@dust-tt/types";
+import type {
+  LightAgentConfigurationType,
+  WorkspaceType,
+} from "@dust-tt/types";
 import { useMemo } from "react";
 
 import { compareAgentsForSort } from "@app/lib/assistant";
+import { useAgentConfigurations } from "@app/lib/swr/assistants";
 
 function makeEditorSuggestions(
   agentConfigurations: LightAgentConfigurationType[]
@@ -17,9 +21,14 @@ function makeEditorSuggestions(
 }
 
 const useAssistantSuggestions = (
-  agentConfigurations: LightAgentConfigurationType[],
-  inListAgentConfigurations: LightAgentConfigurationType[]
+  inListAgentConfigurations: LightAgentConfigurationType[],
+  owner: WorkspaceType
 ) => {
+  const { agentConfigurations } = useAgentConfigurations({
+    workspaceId: owner.sId,
+    agentsGetView: "assistants-search",
+  });
+
   // `useMemo` will ensure that suggestions is only recalculated
   // when `inListAgentConfigurations` or `agentConfigurations` changes.
   const allSuggestions = useMemo(() => {
