+++ front/components/misc/DropzoneContainer.tsx
@@ -1,7 +1,7 @@
 import { DropzoneOverlay } from "@dust-tt/sparkle";
 import { useDropzone } from "react-dropzone";
 
-import { useFileDrop } from "@app/shared/files/FileUploaderContext";
+import { useFileDrop } from "@app/components/assistant/conversation/FileUploaderContext";
 
 interface DropzoneContainerProps {
   children: React.ReactNode;
