--- front/lib/connectors.ts
@@ -1,86 +1,0 @@
-import type { ConnectorProvider } from "@dust-tt/types";
-
-import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
-
-function getConnectorOrder() {
-  return Object.keys(CONNECTOR_CONFIGURATIONS)
-    .filter(
-      (key) =>
-        CONNECTOR_CONFIGURATIONS[key as keyof typeof CONNECTOR_CONFIGURATIONS]
-          .connectorProvider !==
-        CONNECTOR_CONFIGURATIONS.webcrawler.connectorProvider
-    )
-    .map(
-      (key) =>
-        CONNECTOR_CONFIGURATIONS[key as keyof typeof CONNECTOR_CONFIGURATIONS]
-          .connectorProvider
-    );
-}
-
-type ComparableByProvider = { connectorProvider: ConnectorProvider | null };
-function compareByImportance(
-  a: ComparableByProvider,
-  b: ComparableByProvider
-): number {
-  const aConnector = a.connectorProvider;
-  const bConnector = b.connectorProvider;
-
-  const order = getConnectorOrder();
-
-  // Handle null cases.
-  if (aConnector === null) {
-    return bConnector === null ? 0 : 1;
-  }
-  if (bConnector === null) {
-    return -1;
-  }
-
-  // Handle webcrawler cases.
-  if (aConnector === "webcrawler") {
-    return 1;
-  }
-  if (bConnector === "webcrawler") {
-    return -1;
-  }
-
-  // Get indices in sorted connectors.
-  const indexA = order.indexOf(aConnector);
-  const indexB = order.indexOf(bConnector);
-
-  // If both are not found, they are considered equal.
-  if (indexA === -1 && indexB === -1) {
-    return 0;
-  }
-
-  // Compare indices, treating not found as less important.
-  return (
-    (indexA === -1 ? order.length : indexA) -
-    (indexB === -1 ? order.length : indexB)
-  );
-}
-
-// Order in the following format : connectorProvider > empty > webcrawler
-export function orderDatasourceByImportance<Type extends ComparableByProvider>(
-  dataSources: Type[]
-) {
-  return dataSources.sort(compareByImportance);
-}
-
-export function orderDatasourceViewByImportance<
-  Type extends { dataSource: ComparableByProvider },
->(dataSourceViews: Type[]) {
-  return dataSourceViews.sort((a, b) => {
-    return compareByImportance(a.dataSource, b.dataSource);
-  });
-}
-
-export function orderDatasourceViewSelectionConfigurationByImportance<
-  Type extends { dataSourceView: { dataSource: ComparableByProvider } },
->(dataSourceViews: Type[]) {
-  return dataSourceViews.sort((a, b) => {
-    return compareByImportance(
-      a.dataSourceView.dataSource,
-      b.dataSourceView.dataSource
-    );
-  });
-}
